0 warnings

+++ Kernel/syscalls.c
@@ -6,6 +6,7 @@
 #include "drivers/include/keyboard_driver.h"
 #include "keyboard_buffer.h"
 #include "include/interrupts.h"
+#include "drivers/include/ScanCodes.h"
 
 extern int getKey();
 
@@ -36,7 +37,7 @@ void sys_read( char *buf, int len, int filedescriptor){
                     if (aux == 0x39)
                         buf[i++]=' ';
                     else
-                        buf[i++]=ScanCodes[aux];
+                        buf[i++]=ScanCodes[(int)aux];
                     setPos(pos+1);
                 }
                 pos = getBufferPosition();
